+++ ton_client/src/tests/mod.rs
@@ -95,14 +95,12 @@ struct RequestData {
 
 struct TestRuntime {
     pub next_request_id: u32,
-    pub next_callback_id: u32,
     pub requests: HashMap<u32, RequestData>,
 }
 
 impl TestRuntime {
     fn new() -> Self {
         Self {
-            next_callback_id: 1,
             next_request_id: 1,
             requests: HashMap::new(),
         }